[r] Separate integration tests into smaller parts
