+++ tests/integrations/test_instanceof.py
@@ -1,0 +1,13 @@
+from unittest.case import TestCase
+
+from robber import expect
+from tests import must_fail
+
+
+class TestInstanceIntegrations(TestCase):
+    def test_instance(self):
+        expect(expect(None)).to.be.instanceof(expect)
+
+    @must_fail
+    def test_instance(self):
+        expect(expect(None)).to.be.instanceof(TestCase)
