+++ tests/integrations/test_length.py
@@ -1,0 +1,20 @@
+from unittest.case import TestCase
+
+from robber import expect
+from tests import must_fail
+
+
+class TestLengthIntegrations(TestCase):
+    def test_length_success(self):
+        expect([1, 2]).to.have.length(2)
+
+    @must_fail
+    def test_length_success(self):
+        expect([1, 2]).to.have.length(3)
+
+    def test_empty_success(self):
+        expect('').to.be.empty()
+
+    @must_fail
+    def test_empty_failure(self):
+        expect('test').to.be.empty()
